requirements.txt

streamlit>=1.28.0
groq>=0.4.0
minsearch>=0.1.0
python-dotenv>=1.0.0
click>=8.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

